NO-ISSUE moving conversion helpers to separate package
